Use marginal_distribution() instead of marginal_counts() (#1144)
### Summary

This commit updates the Qiskit marginalization function used in
qiskit-experiments. Qiskit has 2 different function available for
marginalizing counts: `marginal_counts()` [1] which is designed to work
with `Counts` or `Result` objects and does **not** respect ordering of
the indices passed in. The other is `marginal_distribution()` [2] which
is designed to work with `Counts`, `Probabilities`,
`QuasiProbabilities`, or a dictionary with string keys and integer or
float values that respects the order of the indices passed in (so it can
be used to permute the bits). In general if you're not working with a
`Result` object it is better to use `marginal_distribution()` because it
is written primarily in Rust and 10-100x faster than
`marginal_counts()`.

### Details and comments

[1]
https://qiskit.org/documentation/stubs/qiskit.result.marginal_counts.html
[2]
https://qiskit.org/documentation/stubs/qiskit.result.marginal_distribution.html