Remove save/load part from T1 tests (#1178)
### Summary

This PR removes the save/load portion of the `TestT1` tests.

### Details and comments

We encounter repeated errors with the save/load part of `TestT1`,
probably resulting from an error in the way the experiment service
handles data. We have not been able to reliably reproduce the problem
yet, and so for now we will move this part of the test into the
`qiskit-ibm-experiment` repository until it is fixed.