Add device component experiment metadata (#1189)
### Summary

This PR adds a ``device_components`` field to the experiment metadata
that is a list of qubits used in the experiment by default. Experiments
that use custom components need to override this metadata field (just
`ResonatorSpectroscopy` so far). It also exposes device components in
the documentation and updates the custom experiment tutorial. #1163
depends on this PR to add the metadata so that non-qubit components can
be correctly identified for figure names.