Avoid setting auto save triggering save for children in a composite experiment (#1187)
### Summary

This PR changes the behavior of `ExprimentData.auto_save` such that
setting this field in the parent of a composite experiment does not
trigger the `save` method in the children experiments.

An unrelated change in this PR is setting a hard coded cap of 10 on the
numbers of workers that can be used in experiment saving.

### Details and comments

Currently setting `auto_save = True` triggers a save of the parent,
including the children. If in addition to that the `auto_save` field is
set to `True` in the children it triggers another save; this is usually
harmless, but it is an unexpected behavior which led to #1184.

This PR changes this. Children will save along with the parents' initial
save. This is done by adding a switch to save (which is on by default
but off when saving from the `auto_save` setter)

---------

Co-authored-by: Helena Zhang <Helena.Zhang@ibm.com>