Fix circuit ordering in InterleavedRB (#1191)
### Summary
Changes the ordering of circuits generated by `InterleavedRB` back to
RIRIRI (R: Reference, I: Interleaved) order. It was accidentally changed
into RRRIII order in #898. Before that, it had been RIRIRI order.

### Details and comments
This change may slightly alter the impact of noise on IRB results.
RIRIRI ordering is preferable in the sense that it minimizes the impact
of drift or noise fluctuations. Note that `InterleavedRBAnalysis` is not
changed that means it works correctly regardless of the circuit ordering
because it processes data relying on "interleaved" flag in circuit
metadata.

This commit also adds a new experiment option `circuit_order` to
`InterleavedRB`. It enables to change the order of the reference and the
interleaved circuits and hence slightly alter the impact of noise on
interleaved RB results. The default value is set to `"RIRIRI"`.

(cherry picked from commit c6753f3be52aaba0b2183120b15845b81ac85d96)