Remove reserved variables (#1202)
### Summary

The variables `__drive_freq_parameter__` and
`__readout_freq_parameter__` have been removed from `Calibrations`.

### Details and comments

- **Why this change was necessary**: These frequency variables were
given special treatment which is inconsistent with the framework.
Furthermore, this way of handling variables that do not have a schedule
was not scalable.

- **What alternative solutions were considered and why the current
solution was chosen**: To replace them a mechanism to define and add
parameters without a schedule has been added to the basis gate library.
This has the benefit of making the API of frequency calibration
experiments consistent with the other calibration experiments.
Furthermore, the proposed change is scalable in that users can add more
parameters that do not have a schedule to a library.

- **What tests and documentation have been added/updated**: A test was
added to make sure that users can add parameters.

- **What do users and developers need to know about this change**: Users
can now add parameters to their library that are not attached to a
schedule in a meaningful way. The variables `__drive_freq_parameter__`
and `__readout_freq_parameter__` have been removed from `Calibrations`.

---------

Co-authored-by: Naoki Kanazawa <nkanazawa1989@gmail.com>