Add coveralls to CI (#1164)
### Summary

This PR reports test coverage to
https://coveralls.io/github/Qiskit-Extensions/qiskit-experiments every
time the CI tests are run. Going forward, the coveralls bot should
report coverage beneath each PR as a comment. This is done by adding a
`cover` environment to tox that replaces the default env in the CI when
running tests + coverage (this avoids having to run tests twice, once
for CI and once to generate coverage). It runs the tests and generates a
coverage report at `coverage.lcov`, which is then sent to coveralls
using its github action.

This PR also adds a temporary version limit to scipy since the newest
1.11 release is currently incompatible with the newest 1.3.1 cvxpy.

### Details

Coverage only runs on MacOS—on Windows, coverage doesn't generate the
report for some reason, and it also seems unnecessary to generate
coverage for multiple OS's.