Use new Furo theme (#1225)
### Summary

This PR switches the docs to the new Furo ecosystem theme.


![image](https://github.com/Qiskit-Extensions/qiskit-experiments/assets/3870315/e93ee925-805e-466a-991a-fd30b663a24f)

### Details and comments

New build variables have been added:
- `FULL_TOCTREE` builds the full API toctree when turned on. By default,
this is off for a fast build time and small file sizes. This is turned
on for deployments. See
https://github.com/Qiskit/qiskit_sphinx_theme/issues/328 for more info.
- `VERSION_STRING` and `RELEASE_STRING` are exposed so that the site can
display an alternative version/release for builds. This is relevant for
the dev site, where we would like to see the `git describe` commit
string instead of the version to be clearer to users. The new theme
currently does not expose the version/release (except for the release in
the page title) since it doesn't have breadcrumbs, but this will be
added after https://github.com/Qiskit/qiskit_sphinx_theme/issues/361.

### Current known issues:

The experiment gallery doesn't render correctly:


![image](https://github.com/Qiskit-Extensions/qiskit-experiments/assets/3870315/13bd6dfc-8732-4337-9e9b-67ce709f7409)