Add optional dependencies file (#1219)
### Summary

This PR adds `requirements-extras.txt` which contains extra packages
that can be installed as `[extras]` and were mostly moved from
`requirements-dev.txt`. These are in contrast to the core dependencies
in `requirements.txt` and the requirements that are only used for tests,
which will stay in `requirements-dev.txt`.

### Details and comments

While the optional packages are not related to each other, it seemed too
complicated to split out optional dependencies into further categories.

There are a few additional changes:

- Added `qiskit-ibm-provider>=0.6.1` as an optional dependency since
this release adds numpy serialization, which fixes some known bugs with
circuit metadata serialization (see also #1176)
- Unpinned `scipy` (done in #1164) and bumped `cvxpy` to 1.3.2, which is
now compatible with scipy 1.11
- Removed matrix variables from CI workflow that doesn't use them
- Updated install and contributing instructions