Bugfix in ExperimentData's handling of metadata save (#1229)
### Summary

This PR fixes a bug resulting in experiment data saving fails due to
large payloads.

### Details and comments

Since the `metadata` field of `ExperimentData` can grow tremendously
large, resulting in a save failure due to the large payload. we have
added has the ability to save the metadata as an artifact file. However,
a small bug prevented this from triggering. This PR fixes the bug and
adds a relevant test.

---------

Co-authored-by: Helena Zhang <Helena.Zhang@ibm.com>