Speed up non-production docs builds (#1237)
### Summary

This PR turns off source code viewing for non-production builds to speed
them up (the speed up seems to be only ~2 minutes, but better than
nothing). It also turns on parallel builds for CI which was disabled
because of random crashes on MacOS, but that shouldn't be a problem for
CI since it's on Ubuntu (this speed up is around ~8 minutes). The
`FULL_TOCTREE` flag has been changed to `PROD_BUILD` to encompass all
the changes needed for a production vs. non-production build.

In addition, Getting Started has been removed from the tutorials index
so that it doesn't show up twice on the sidebar. Some typos have been
fixed.