Bugfix in hgp setter in `ExperimentData` (#1234)
### Summary
This PR fixes a bug in :meth:`.ExperimentData.hgp`.

### Details and comments
Currently :meth:`.ExperimentData.hgp` uses a regex which is too strict
and cannot recognize strings such as "ibm-q-internal". This PR updates
the regex to accept everything of the form "A/B/C" where A,B,C do not
contain "/" but can contain anything else.