Introduce dataframe to ExperimentData (step1) (#1133)
### Summary

Executing approved design proposal in
https://github.com/Qiskit/rfcs/blob/master/0007-experiment-dataframe.md.
Introduction of the `Artifact` will be done in the followup.

This PR introduces dataframe to `ExperimentData` and replaces the
conventional
[AnalysisResult](https://github.com/Qiskit/qiskit-experiments/blob/main/qiskit_experiments/framework/analysis_result.py)
with it. With this PR, experimentalist can visualize all analysis
results with a single html table by `dataframe=True`.


![image](https://user-images.githubusercontent.com/39517270/230959307-d5a4d471-1659-4cf4-974e-cdb21e625ad8.png)

Once can control the columns to show in the html table with `columns`
option.


### Details and comments

Experiment analysis is a local operation on client side, and data
generated there must be agnostic to the payload format of remote
service. In the current implementation, `AnalysisResult` appears in the
`BaseAnalysis` and the data generated by the analysis is implicitly
typecasted to this class. Note that `AnalysisResult` *not only* defines
canonical data format *but also* implements the API for the IBM
Experiment service. This tight coupling to the IBM Experiment service
may limit data expressibility and an analysis class author always need
to be aware of the data structure of the remote service.

In this PR, internal data structure is replaced with the flat dataframe,
and one should be able to add arbitrary key-value pair to the analysis
result without concerning about data model. Added key creates a new
column in the table, which must be internally formatted and absorbed by
extra field of the payload at a later time when these results are saved.
This is implemented by
`qiskit_experiments.framework.experiment_data._series_to_service_result`.
Bypassing the generation of `AnalysisResult` object also drastically
improves the performance of heavily parallelized analysis that generates
huge amount of data.

---------

Co-authored-by: Helena Zhang <Helena.Zhang@ibm.com>