Remove reset from beginning of tomography experiments (#1250)
### Summary

This change removes the reset instruction from the beginning of
tomography experiments.

### Details and comments

Some backends do not implement reset but can be treated as having reset
the qubits between circuits. It is also possible some backends implement
reset but have a slower, higher fidelity reset between circuits than the
in-circuit reset.