Fix calculation of curve fit weights (#1224)
### Summary

This PR updates calculation of weights to compute residual in the curve
fitting.

### Details and comments

When the error bar of data points is significantly small, these data
points become a dominant source of residual to minimize. This means
other data points contribute little to the fit, and causes local overfit
to certain data points. This is fixed by clipping the weights to remove
outlier.