Address timing dependence in ExperimentData.add_jobs test (#1267)
Previously, the test `test_add_data_job` called `add_jobs` with two jobs
and then tested that the results from `ExperimentData.data()` matched
the results of the first job concatenated with the second.
`ExperimentData` uses threads to handle tracking jobs. Even though these
jobs are dummy jobs that take no time, `ExperimentData` would
occasionally add the results from the second job before the first and
then fail the results comparison. With this change, a label is attached
to each result and then the results are sorted based on the label before
testing against the expected results.

Closes #1044