Faster TestEPGAnalysis (#1276)
### Summary
Made `TestEPGAnalysis` faster so that they can run within the 60 second
limit.

### Details and comments
The heavy part in the analysis tests is in running experiments before
the analyses to be tested. So common experiments are run in advance and
used for analyses in `TestEPGAnalysis`. However, all the common
experiments were run before every test because they were run in `setUp`.
This commit moves them to `setUpClass` so that they are run once as
expected. By the change, the total execution time is reduced from 34 sec
to 7 sec in my local environment.