Figure extraction from experiment data by name without extension (#1287)
### Summary

Figure names could be referred without ".svg" at the ending.
### Details and comments

Some details that should be in this section include:
In #629 , an issue was raised where a figure that was saved with the
name "abcd" couldn't be accessed. This issue is caused by that the
'add_figures()' method in `ExperimentData` class added ".svg"
automatically. As it explicitly written that SVG extension is added if
the file name doesn't have one, we can conclude that all figure names
have ".svg" extension. To not make a major change in the code I added to
the `figure` method in `ExperimentData` class a check for the extension
of the `figure_key`. If it doesn't have ".svg" ending, it adds one
before trying to pull it from the dict of figures.

---------

Co-authored-by: Yael Ben-Haim <yaelbh@il.ibm.com>