Change RB test parameters to make the test faster and more robust
`test_add_more_circuit_yields_lower_variance` started failing with
qiskit-aer 0.13.0. The test was only comparing the results of 3 RB
sampels to 5 RB samples and looking for the case with more samples to
have lower variance. These numbers of samples were likely not enough to
ensure that the case with more samples always had lower variance across
a wide range of random seeds. Here the high number of samples was
increased to 30.

To make sure that the fit results were reasonable and
that the test run time did not get too long, these additional changes
were:

* Convert the test from two qubit RB to single qubit RB. This makes the
  simulation faster.
* Increase the single qubit gate error. This makes the qubit depolarize
  faster so that the RB curve can be traced out with less gates.