Update docstring options parser and add docs cron job (#1312)
Qiskit 0.45 introduces singleton gates, which have dynamically generated
type objects, so their associated module have to be accessed from the
`base_class` attribute. This PR update the docstring options parser to
work correctly with these objects. It also specifies the `iqp` style
when drawing circuits to avoid deprecation warnings during the build,
and adds the docs build to the daily cron job so this kind of issue can
be caught more quickly in the future.

Because Aer 0.13.0 is causing a slowdown in docs builds, this PR also
temporarily pins the version to 0.12.2 until the next patch release.