Raising a warning for running analysis on an empty `ExperimentData` object (#1313)
### Summary

Added a warning for the user if the `ExperimentData` data is empty
before running an analysis and if the data couldn't be retrieved.

### Details and comments

Some details that should be in this section include:

- The user will get an un-inductive error regarding the analysis while
the reason it was raised is because the analysis doesn't have data to
analyze from the experiment. We can see an example in #1182 .
- A solution that was suggested was to raise an error that says the user
is trying to run analysis on empty `ExperimentData` object but the team
didn't want to constraint the user to run analysis only if the
experiment has data so we decided that raising a warning is enough.

Changes:
- [x] Added warning before analysis if `ExperimentData.data()` is empty.

### PR checklist (delete when all criteria are met)

- [x] I have read the contributing guide `CONTRIBUTING.md`.
- [x] I have added a release note file using `reno` if this change needs
to be documented in the release notes.

---------

Co-authored-by: Yael Ben-Haim <yaelbh@il.ibm.com>