Replace qiskit-terra dependency with qiskit dependency
This change brings qiskit-experiments in line with other packages like
qiskit-ibm-provider and qiskit-aer and prepares for the deprecation of
the qiskit-terra package.

The effective minimum version of qiskit-terra is bumped up one minor
version from 0.24 to 0.25 (linked to qiskit 0.44) as part of this
change, but since qiskit 0.45 has already been released this should not
be an impactful change for most users.