Replace qiskit-terra references with qiskit references (#1318)
This change brings qiskit-experiments in line with other packages like
qiskit-ibm-provider and qiskit-aer and prepares for the deprecation of
the qiskit-terra package.

Detailed changes:

* Replace qiskit-terra dependency with qiskit dependency
- The effective minimum version of qiskit-terra is bumped up one minor
version from 0.24 to 0.25 (linked to qiskit 0.44) as part of this
change, but since qiskit 0.45 has already been released this should not
be an impactful change for most users.
* Simplify the code for logging the version of qiskit used. It should
always use the Qiskit version as a string. Previously, it could
sometimes use a dictionary with the Qiskit and Qiskit Experiments
versions.
* Remove workaround for `aquire_alignment` misspelling in old versions
of terra (because it was commented referencing terra)
* Replace qiskit-terra in requirements-dev.txt and tox.ini
* Replace various references to "terra" in various code comments with
"Qiskit"
* Rename terra-main and docs-terra-main tox environments to qiskit-main
and docs-qiskit-main
* Rename GitHub workflow job terra-main-tests to qiskit-main-tests

The only remaining references to "terra" in the codebase are in old
release notes.