Bugfix when copying the `ExperiemntData` object and add warning to `_retrieve_data` (#1316)
### Summary

As said in #1314 , `copy` method in `Experiment Data` class will now
copy the `provider` attribute to the copied instance and warnning if
data isn't retrieved due to `provider=None` .
### Details and comments

Some details that should be in this section include:

- Why this change was necessary
This is a bug that the provider isn't copied. In addition,
`_retrive_data()` method in the `ExperimentData` class doesn;t retirive
data if provider is not provided. Added a warning so the user will know.