Reduce default parallelism of tox test runs
Measurements of the total test time for different parallelism settings
showed that total test time increased when `QISKIT_PARALLEL` was set to
`TRUE` or `OMP_NUM_THREADS` was set higher than 1 when the `stestr`
concurrency was higher than the number of physical cores. Additionally,
no improvement was seen when increasing the `stestr` concurrency beyond
the number of physical cores (for example up to two times that number on
a system with 2x multithreading), but also no degradation in performance
was observed, so the `stestr` concurrency was left at the default value
(the number of logical CPUs rather than the number of physical CPUs) for
simplicity.