Suppress more numpy warnings from uncertainties (#1350)
This PR is a follow up to
https://github.com/Qiskit-Extensions/qiskit-experiments/pull/1070. It
suppresses numpy warnings during uncertainties array creation in more
places in the code. Something changed recently in the numpy code so that
it generates warnings when it used to suppress them, in particular when
using `numpy.vectorize` which wraps user-level Python code inside of
numpy C code and seems to lose context about warning state. See
https://github.com/numpy/numpy/issues/21416 for more information.