Use tox plugin to consolidate tox environments
This change adds a tox plugin in `toxfile.py` which allows for sets of
commands to be mapped to names in the test environment configuration.
Additionally, a command line option is added to `tox` which allows one
of these command sets to be selected by name and run in place of the
default `commands` entry. With this alternative commands feature,
several of the test environments that were in `tox.ini` were
consolidated into three remaining environments with different sets of
commands. The remaining environments are:

* The main Python test environment used for tests, linting, etc.
* The docs environment, needed because of a Sphinx bug that prevents a
  development install of `qiskit-experiments` from working on macOS.
* The test environment that uses the Qiskit `main` branch instead of the
  released package.