Remove unnecessary circuit metadata (#1315)
### Summary

As the number of qubits in the parallel experiment increases, the memory
footprint of job payload communicated through wire becomes more serious
issue. Some built-in experiment generates circuits with unnecessary
metadata, which is
- duplicated information existing in experiment metadata
- duplicated information existing in every circuit metadata
- un-used information in analysis

This PR removes such unnecessary fields from the circuit metadata to
reduce payload size. For example, if an experiment is paired with a
curve analysis, it only requires `xval` and other minimum keys necessary
for model matching.

### Details and comments

Tutorial is also updated.

---------

Co-authored-by: Will Shanks <wshaos@posteo.net>
Co-authored-by: Helena Zhang <Helena.Zhang@ibm.com>