Set matplotlib axis scale type earlier so axis limits are correct
When the axis scale is changed to "log", matplotlib changes the default
axis limits from (0, 1) to something around (0.1, 10). Since MplDrawer
got the limits and the set them back on the axes after setting the
scale, it was reapplying a minimum of 0 to the log scale plot, leading
to a user warning. matplotlib by default changes the limits to the range
of the data when data is added, so this case only came up when
generating a figure with no data.