Fix guess.rb_decay not to raise an error against bad output (#1336)
### Summary
Fixed a bug in `guess.rb_decay` where it unintentionally raises an error
if all y values are below the b value (that can happen in experiments on
real devices).

### Details and comments
Changed to return `0` as the initial guess of the decay parameter when
all y values are below the b value, instead of raising an IndexError
like
```
File "/qiskit-experiments/qiskit_experiments/curve_analysis/guess.py", line 406, in rb_decay
    return ((y[0] - b) / a) ** (1 / x[0])
IndexError: index 0 is out of bounds for axis 0 with size 0
```

An alternative solution might be explicitly raising a QiskitError. But,
if we want to raise an error for too bad data to fit, I think the fitter
(not rb_decay) should raise an error because it should not be due to the
failure of guessing initial parameter (that is the object of `rb_decay`
function.

---------

Co-authored-by: Naoki Kanazawa <nkanazawa1989@gmail.com>